+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+* [CLI] Add tracing logs in all layers.
 * [CLI] Add a new global `-v...` flag to control the log verbosity. By default,
     no logs are emitted.
 * [`git z commit`] Ask whether to reuse the previous answers or commit message